+++ localflavor/cu/forms.py
@@ -99,7 +99,7 @@ class CUPostalCodeField(RegexField):
     def to_python(self, value):
         value = super().to_python(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         return value.strip()
 
 
@@ -134,5 +134,5 @@ class CUIdentityCardNumberField(RegexField):
     def to_python(self, value):
         value = super().to_python(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         return value.strip()