+++ localflavor/cz/forms.py
@@ -42,7 +42,7 @@ class CZPostalCodeField(RegexField):
         """
         value = super().clean(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         return value.replace(' ', '')