+++ localflavor/np/forms.py
@@ -23,33 +23,30 @@ class NPPostalCodeFormField(RegexField):
     def __init__(self, **kwargs):
         super().__init__(r'^\d{5}$', **kwargs)
 
-    def clean(self, value):
-        value = super().clean(value)
-        if value in self.empty_values:
-            return self.empty_value
-        return value
 
 class NPZoneSelect(Select):
     """
-    A Select widget with option to select a zone from 
+    A Select widget with option to select a zone from
     list of all zones of Nepal.
     """
 
     def __init__(self, attrs=None):
         super().__init__(attrs, choices=ZONES)
 
+
 class NPProvinceSelect(Select):
     """
-    A Select widget with option to select a province from 
+    A Select widget with option to select a province from
     list of all provinces of Nepal.
     """
 
     def __init__(self, attrs=None):
         super().__init__(attrs, choices=PROVINCES)
 
+
 class NPDistrictSelect(Select):
     """
-    A Select widget with option to select a district from 
+    A Select widget with option to select a district from
     list of all districts of Nepal.
     """