+++ localflavor/pe/forms.py
@@ -29,7 +29,7 @@ class PEDNIField(CharField):
         """Value must be a string in the XXXXXXXX formats."""
         value = super().clean(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         if not value.isdigit():
             raise ValidationError(self.error_messages['invalid'], code='invalid')
         if len(value) != 8:
@@ -57,7 +57,7 @@ class PERUCField(CharField):
         """Value must be an 11-digit number."""
         value = super().clean(value)
         if value in self.empty_values:
-            return self.empty_value
+            return value
         if not value.isdigit():
             raise ValidationError(self.error_messages['invalid'], code='invalid')
         if len(value) != 11: