+++ localflavor/za/forms.py
@@ -26,7 +26,7 @@ class ZAIDField(CharField):
         value = super().clean(value)
 
         if value in self.empty_values:
-            return self.empty_value
+            return value
 
         # strip spaces and dashes
         value = value.replace(' ', '').replace('-', '')